fix(go): regenerate corrupt exchange_generated.go base
The BaseExchange tier-split migration hand-edited the generated Go base and left
exchange_generated.go with botched method splices (a FetchTicker header on a
PriceToPrecision body) plus dangling Options-struct references in the typed
interface, so go build ./v4/ failed to parse. Regenerated cleanly from the TS
source; go/v4 and go/v4/prediction now build.

+++ go/v4/lighter_wrapper.go
@@ -44,7 +44,7 @@ func (this *Lighter) CreateSubAccount(name string, options ...CreateSubAccountOp
 	}
 	return res.(map[string]any), nil
 }
-func (this *Lighter) FetchNonce(accountIndex any, apiKeyIndex any, options ...FetchNonceOptions) (int64, error) {
+func (this *Lighter) FetchNonce(accountIndex any, apiKeyIndex any, options ...FetchNonceOptions) (float64, error) {
 
 	opts := FetchNonceOptionsStruct{}
 
@@ -58,9 +58,9 @@ func (this *Lighter) FetchNonce(accountIndex any, apiKeyIndex any, options ...Fe
 	}
 	res := <-this.Core.FetchNonce(accountIndex, apiKeyIndex, params)
 	if IsError(res) {
-		return -1, CreateReturnError(res)
+		return float64(-1), CreateReturnError(res)
 	}
-	return (res).(int64), nil
+	return (res).(float64), nil
 }
 
 /**
